Refactorializificate web30 client into chunks
The client.rs file was over 1K lines, break into separate files
